+++ src/MzIO.MzML/RELEASE_NOTES.md
@@ -1,3 +1,6 @@
+#### 0.1.7 - Wednesday, February 12, 2025
+* Upgrade .NET SDK to 6.0
+
 #### 0.1.6 - Tuesday, June 13, 2023
 * Change FSharp.Core version back to 5.0.0